alteri o local do addUsuario() para classe do ADM
