+++ Sistema.java
@@ -5,7 +5,7 @@ import java.util.LinkedList;
 import java.util.Scanner;
 
 public class Sistema {
-    private LinkedList<Usuario> lstUsuarios;
+    static LinkedList<Usuario> lstUsuarios;
     Scanner in = new Scanner(System.in);
     Scanner inS = new Scanner(System.in);
     private int qtdMedicos;
@@ -227,37 +227,7 @@ public class Sistema {
         return null;
     }
 
-    public void addUsuario(){
-        boolean aux = false;
-        int id = 0;
-        do{ //VALIDACAO PARA NAO TER IDS IGUAIS 
-            System.out.println("Qual o id do usuario? ");
-            id = in.nextInt();
-
-            for (Usuario lst : lstUsuarios) {
-                if (lst.getId()!=id) {
-                    aux = true;
-                }
-            }
-        }while(aux==false);
-
-        System.out.println("Qual o nome do usuario? ");
-        String nome = inS.nextLine();
-        System.out.println("Qual o tipo do usuario? ");
-        String tipo = inS.nextLine();
-
-        if(tipo.equalsIgnoreCase("Paciente")){ //ADD AS INFOS NA LINKEDLIST
-            Usuario u = new Paciente(id, nome, tipo);
-            lstUsuarios.add(u);
-        } else if(tipo.equalsIgnoreCase("Medico")){
-            Usuario u = new Medico(id, nome, tipo);
-            lstUsuarios.add(u);
-        } else if(tipo.equalsIgnoreCase("Administrador")){
-            Usuario u = new Administrador(id, nome, tipo);
-            lstUsuarios.add(u);
-        }
-    }
-
+    
     public String buscaUsuario(){ //DEPENDENDO DE COMO FICAR A AUTORIZACAO VAI TER QUE MUDAR A SAIDA DO METODO
         System.out.println("Qual o nome do usuario a ser pesquisado? ");
         String nome = inS.nextLine();
